Split code into smaller components
